Add functionality to raise exception if possibly stalled  (#16)
There are two changes here. The first is adding a function to replace tick for non-blocking state machines which raises an exception if the state machine finishes without completing. If also runs the state machine while supporting retry counts for non-blocking state machines which just calling a single tick will not support.

The other change is to raise an exception in blocking state machine (in tick) if the state we are blocking in doesn't have a timeout value set and we could stall indefinitely. This change does break current kiosk software. As shown in the event_driven.py example, untimed states can still be used but checked to make sure you only have expected indefinite states.